Support older (<5.6) kernels
openat2 was introduced in 5.6 so hooking it
fails on older kernels. This fixes that by
making it optional.

Older kernels also have trouble with .rodata in BPF.
The exact cause is unclear but seems like an interaction
of libbpf and the kernel. To work aroud that, don't printk
by default (uncomment to use it to debug). That's b/c
string literals introduce .rodata section.