gh: support SBOM upload from a forked repo
If a pull request is submitted from a forked PR,
it does not have access to secrets (EDGEBIT_TOKEN).

To workaround this, the PR triggered workflow generates an
SBOM and uploads it as an artifact (along with some metadata).

Another workflow gets kicked off once the first one is done.
It uploads the SBOM to EdgeBit.