Log at info (vs error) if can't save base-os ID
On k8s deployments, there's often no volume to persist
the BaseOS generated UUID. An info level log is more
appropriate.

If the BaseOS UUID is not saved, it'll be regenerated on
agent restart. The EdgeBit platform deals with it ok, there's
a brief period of time when a machine will show multiple workloads.